pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "rabbitviewer"
version = "0.1.0"
description = "A daemon + GUI image viewer"
requires-python = ">=3.10"
dependencies = [
    "PySide6>=6.8.2,<7",
    "PyYAML>=6.0",
    "Pillow",
    "watchdog>=2.1.9",
    "send2trash",
    "pydantic>=2.0",
]

[project.optional-dependencies]
cr3 = ["rawpy", "numpy", "imageio"]
dev = ["pytest==7.2.1", "pytest-qt==4.0.2"]

[project.scripts]
rabbitviewer = "main:main"
rabbitviewer-daemon = "rabbitviewer_daemon:main"

[tool.setuptools]
py-modules = ["main", "rabbitviewer_daemon"]

[tool.setuptools.packages.find]
exclude = ["tests*", "venv*", "benchmarks*", "*.egg-info"]

[tool.setuptools.package-data]
"*" = ["*.png", "*.svg", "*.webp"]
